move gitProvider interface to graphql package
